Former-commit-id: 1125e4948a1c7bd300316be0ebeef34fde3b1a2b

+++ app/Http/Controllers/ContractsController.php
@@ -464,7 +464,7 @@ class ContractsController extends Controller
 
         $request->session()->flash('message.nivel', 'success');
         $request->session()->flash('message.content', 'Your contract was updated');
-        return redirect()->route('RequestImportation.index');
+        return redirect()->route('RequestFcl.index');
     }
 
     public function show($id)
@@ -845,7 +845,7 @@ class ContractsController extends Controller
 
         $request->session()->flash('message.nivel', 'success');
         $request->session()->flash('message.title', 'Well done!');
-        $request->session()->flash('message.content', 'You successfully update this contract.');
+        $request->session()->flash('message.content', 'The contract was successfully updated');
         return redirect()->back()->with('editContract','true');
 
         //return redirect()->action('ContractsController@index');
@@ -1150,9 +1150,13 @@ class ContractsController extends Controller
             }
 
             $contract = Contract::find($id);
-            $contract->delete();
+            if(!empty($contract)){
+              $contract->delete();
+              return response()->json(['message' => 'Ok']);
+            }else{
+              return response()->json(['message' => 'HasDeleted']);
+            }
 
-            return response()->json(['message' => 'Ok']);
         }
         catch (\Exception $e) {
             return response()->json(['message' => $e]);
@@ -1488,16 +1492,21 @@ class ContractsController extends Controller
         $carrier        = Carrier::pluck('name','id');
         $directions     = Direction::pluck('name','id');
         $contract       = Contract::find($id);
+        if(!empty($contract->gp_container_id)){
+            $equiment_id = $contract->gp_container_id;
+        } else {
+            $equiment_id = 1;
+        }
         $contract->load('carriers');
         $companyUsers   = CompanyUser::pluck('name','id');
         //dd($companyUser);
-        return view('contracts.Body-Modals.DuplicatedscontractsOtherCompany',compact('contract','carrier','directions','companyUsers','request_dp_id'));
+        return view('contracts.Body-Modals.DuplicatedscontractsOtherCompany',compact('contract','carrier','directions','companyUsers','request_dp_id','equiment_id'));
     }
 
     public function duplicatedContractStore(Request $request, $id){
         $requestArray   = $request->all();
         $requestArray['requestChange'] = false;
-        $data           = ['id'=> $id,'data' => $requestArray];
+        $data = ['id'=> $id,'data' => $requestArray];
         if(env('APP_VIEW') == 'operaciones') {
             GeneralJob::dispatch('duplicated_fcl',$data)->onQueue('operaciones');
         } else {
@@ -1505,13 +1514,13 @@ class ContractsController extends Controller
         }
 
         $request->session()->flash('message.nivel', 'success');
-        $request->session()->flash('message.content', 'The contract is duplicating, please do not delete it, is doubling');
+        $request->session()->flash('message.content', 'The contract is duplicating, please do not delete it');
         return redirect()->route('contracts.index');
     }
 
     public function duplicatedContractFromRequestStore(Request $request, $id){
         $requestArray   = $request->all();
-        //dd($requestArray);
+        
         $requestArray['requestChange'] 	= true;
         $time   = new \DateTime();
         $now2   = $time->format('Y-m-d H:i:s');
@@ -1541,7 +1550,7 @@ class ContractsController extends Controller
         }
 
         $request->session()->flash('message.nivel', 'success');
-        $request->session()->flash('message.content', 'The contract is duplicating, please do not delete it, is doubling');
+        $request->session()->flash('message.content', 'The contract is duplicating, please do not delete it');
         return redirect()->back();
     }
